Refactor runner utils and add tests
